boxoffice/requirement.txt
appdirs
bcrypt
cffi
django-audit-log
django-auditlog
django-jsonfield
django-sslserver
packaging
pycparser
pyparsing
pytz
six